Change quantity and price type

+++ cxx/ITradeclient.h
@@ -20,7 +20,7 @@ public:
   virtual auto stop() const -> void = 0;
 
   virtual auto put_order(const std::string &symbol, const char side,
-                         const uint32_t quantity, const uint32_t price,
+                         const double quantity, const double price,
                          const char time_in_force) const
       -> std::unique_ptr<std::string> = 0;
   virtual auto cancel_order(const std::string &order_id,
